24w21b

+++ src/main/java/me/deftware/mixin/mixins/render/MixinCamera.java
@@ -42,7 +42,7 @@ public class MixinCamera implements GameCamera {
     private final EventCameraClip eventCameraClip = new EventCameraClip();
 
     @Inject(at = @At("HEAD"), cancellable = true, method = "clipToSpace")
-    public void clipToSpace(double camDistance, CallbackInfoReturnable<Double> info) {
+    public void clipToSpace(float camDistance, CallbackInfoReturnable<Double> info) {
         eventCameraClip.create(camDistance);
         eventCameraClip.broadcast();
         if (eventCameraClip.isCanceled()) {
